Revert "The jarjar jars were too big. Go back to regular jars."
This reverts commit fd5eeb401243cfe4db5fba5e35ef0d414b03274d.

Eclipse is forcing my hand. And I really don't want to re-roll bindgen everytime
exigence-utilities changes just to update the classes it is squatting on.

As a compromise, besides rewriting the package name, we also now at a 'Bindgen'
prefix, which will help with the rewritten classes not annoyingly showing up in
Ctrl-Shift-T results.